Start converting web-ui in backend/frontend. Moved 'server' package to 'backend'. Added flask-restufl to requirements.txt.
