+++ code/web/backend/static/js/includes/jquery.cookie.js
