+++ code/web/backend/static/js/svg-pan-zoom.js
