+++ code/web/wsgi.py
@@ -2,7 +2,7 @@ from utils.setup_logging import setup_logging
 
 setup_logging()
 
-from web.server import create_app as game_app
+from web.backend import create_app as game_app
 
 application = game_app()
